fix /discover build
