+++ app/discover/page.tsx
@@ -1,5 +1,4 @@
 import { getTranslations } from "next-intl/server";
-// import { SearchParams } from "nuqs";
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 import { Suspense } from "react";
 import { DiscoverEventsList } from "./discover-events-list";
@@ -7,7 +6,6 @@ import { getQueryClient } from "@/lib/get-query-client";
 import { ScreenContainer } from "@/components/layout/ScreenContainer";
 import { eventsList } from "@/lib/queries/events-list";
 import { EventsListLayout } from "@/components/layout/events-list-layout";
-// import { loadEventFilterSearchParams } from "@/lib/searchParams";
 
 export const revalidate = 60;
 export const dynamicParams = true;
@@ -16,35 +14,25 @@ export async function generateStaticParams() {
   return [];
 }
 
-// type PageProps = {
-//   searchParams: Promise<SearchParams>;
-// };
-
 export default async function DiscoverPage() {
-  // { searchParams }: PageProps
-  // const { from } = await loadEventFilterSearchParams(searchParams);
   const queryClient = getQueryClient();
   const now = Date.now() / 1000;
 
+  // Prefetch upcoming
   queryClient.prefetchQuery(eventsList(now, Number.MAX_SAFE_INTEGER, 20));
+  // Prefetch past
   queryClient.prefetchQuery(eventsList(now - 1, 0, 20));
-  // switch (from) {
-  //   case "upcoming":
-  //     break;
-  //   case "past":
-  //     break;
-  // }
 
   const t = await getTranslations("discover");
 
   return (
     <HydrationBoundary state={dehydrate(queryClient)}>
       <ScreenContainer>
-        <Suspense>
-          <EventsListLayout title={t("title")} description={t("description")}>
+        <EventsListLayout title={t("title")} description={t("description")}>
+          <Suspense>
             <DiscoverEventsList now={now} />
-          </EventsListLayout>
-        </Suspense>
+          </Suspense>
+        </EventsListLayout>
       </ScreenContainer>
     </HydrationBoundary>
   );
