+++ components/widgets/header-events-list.tsx
@@ -1,24 +1,13 @@
-"use client";
-
-import { useTranslations } from "next-intl";
-import { parseAsStringLiteral, useQueryState } from "nuqs";
+import { Suspense } from "react";
 import Heading from "../texts/heading";
 import { GnowebButton } from "../buttons/GnowebButton";
 import Text from "../texts/text";
-import { Tabs, TabsList, TabsTrigger } from "@/components/shadcn/tabs";
+import FromFilterTab from "./from-filter-tab";
 
 const HeaderEventsList: React.FC<{
   title: string;
   description?: string;
 }> = ({ title, description }) => {
-  const [tab, setTab] = useQueryState<"upcoming" | "past">(
-    "from",
-    parseAsStringLiteral(["upcoming", "past"] as const)
-      .withDefault("upcoming")
-      .withOptions({ shallow: false, throttleMs: 200 }),
-  );
-  const t = useTranslations("events-list");
-
   return (
     <div className="flex flex-col gap-2 mb-3">
       <div className="flex flex-col gap-2 md:gap-0 md:flex-row md:justify-between md:items-center">
@@ -31,15 +20,9 @@ const HeaderEventsList: React.FC<{
           />
         </div>
         <div>
-          <Tabs
-            value={tab}
-            onValueChange={(value) => setTab(value as "upcoming" | "past")}
-          >
-            <TabsList className="grid w-full grid-cols-2">
-              <TabsTrigger value="upcoming">{t("upcoming")}</TabsTrigger>
-              <TabsTrigger value="past">{t("past")}</TabsTrigger>
-            </TabsList>
-          </Tabs>
+          <Suspense>
+            <FromFilterTab />
+          </Suspense>
         </div>
       </div>
